+++ node/src/controllers/UsuarioController.js
@@ -1,5 +1,6 @@
 const Sequelize = require('sequelize');
 const Usuario = require('../models/Usuario');
+const Endereco = require('../models/Endereco');
 
 module.exports = {
   async findAll(req, res) {
@@ -9,6 +10,48 @@ module.exports = {
         attributes: ['codigo', 'nome', 'email', 'celular', 'dataNascimento', 'dataCadastro', 'dataAlteracao', 'nivel', 'idEndereco', 'idFotoPerfil'],
         offset: (page - 1) * 5,
         limit: 5
+    })
+    .then( usuarios => {
+        res.header('X-Total-Count', usuarios.count);
+
+        return res.status(200).json(usuarios.rows);
+    })
+    .catch( error => {	
+        return res.status(500).json(error);	
+    });
+
+  },
+  async findAllUsers(req, res) {
+    const { page = 1 } = req.query;
+
+    await Usuario.findAndCountAll({
+        attributes: ['codigo', 'nome', 'email', 'celular', 'dataNascimento', 'dataCadastro', 'dataAlteracao', 'nivel', 'idEndereco', 'idFotoPerfil'],
+        where: {
+            nivel: 'USU'
+        },
+        offset: (page - 1) * 5,
+        limit: 5,
+    })
+    .then( usuarios => {
+        res.header('X-Total-Count', usuarios.count);
+
+        return res.status(200).json(usuarios.rows);
+    })
+    .catch( error => {	
+        return res.status(500).json(error);	
+    });
+
+  },
+  async findAllOrgs(req, res) {
+    const { page = 1 } = req.query;
+
+    await Usuario.findAndCountAll({
+        attributes: ['codigo', 'nome', 'email', 'celular', 'dataNascimento', 'dataCadastro', 'dataAlteracao', 'nivel', 'idEndereco', 'idFotoPerfil'],
+        where: {
+            nivel: 'ORG'
+        },
+        offset: (page - 1) * 5,
+        limit: 5,
     })
     .then( usuarios => {
         res.header('X-Total-Count', usuarios.count);
@@ -37,6 +80,30 @@ module.exports = {
         return res.status(500).json(error);	
     });
   },
+  async Search(req, res) {
+    const { pesquisa } = req.params;
+    const { page = 1 } = req.query;
+
+    await Usuario.findAndCountAll({
+        attributes: ['codigo', 'nome', 'email', 'celular', 'dataNascimento', 'dataCadastro', 'dataAlteracao', 'nivel', 'idEndereco', 'idFotoPerfil'],
+        where: {
+            nome: {
+                [Sequelize.Op.like]: `%${pesquisa}%`
+            }
+        },
+        offset: (page - 1) * 5,
+        limit: 5,
+    })
+    .then( usuarios => {
+        res.header('X-Total-Count', usuarios.count);
+
+        return res.status(200).json(usuarios.rows);
+    })
+    .catch( error => {	
+        return res.status(500).json(error);	
+    });
+
+  },
   async profile(req,res){
       const { codigo } = req.params;
 
@@ -57,6 +124,10 @@ module.exports = {
   async storeUser(req, res) {
     const { nome, senha, email, celular, dataNascimento, cpf } = req.body;
 
+    if(!dataNascimento || dataNascimento == null || dataNascimento == undefined){
+        return res.status(400).json({error: { message: 'Insira a data de nascimento'}});
+    }
+    
     await Usuario.create({
         nome, senha, email, nivel: 'USU' ,  celular, dataNascimento, cpf
     })
@@ -72,10 +143,10 @@ module.exports = {
 
   },
   async storeOrg(req, res) {
-    const { nome, senha, email, celular, dataNascimento, cnpj } = req.body;
+    const { nome, senha, email, celular, cnpj } = req.body;
 
     await Usuario.create({
-        nome, senha, email, nivel: 'ORG', celular, dataNascimento, cnpj
+        nome, senha, email, nivel: 'ORG', celular, cnpj
     })
     .then(usuario => {
         return res.status(201).json(usuario);
@@ -96,12 +167,14 @@ module.exports = {
         nome, senha, email, celular, dataNascimento, cpf
     },{
         where: {
-            codigo
+            codigo, nivel: 'USU'
         }
     })
     .then(async retorno => {
         if(retorno >= 1){
-            await Usuario.findByPk(codigo)
+            await Usuario.findByPk(codigo, {
+                attributes: ['codigo', 'nome', 'email', 'celular', 'dataNascimento', 'dataCadastro', 'dataAlteracao', 'nivel', 'cpf', 'idEndereco', 'idFotoPerfil']
+            })
             .then(usuario => {
                 return res.status(200).json({
                     usuario,
@@ -128,12 +201,14 @@ module.exports = {
         nome, senha, email, celular, dataNascimento, cnpj
     },{
         where: {
-            codigo
+            codigo, nivel: 'ORG'
         }
     })
     .then(async retorno => {
         if(retorno >= 1){
-            await Usuario.findByPk(codigo)
+            await Usuario.findByPk(codigo, {
+                attributes: ['codigo', 'nome', 'email', 'celular', 'dataNascimento', 'dataCadastro', 'dataAlteracao', 'nivel', 'cnpj', 'idEndereco', 'idFotoPerfil']
+            })
             .then(usuario => {
                 return res.status(200).json({
                     usuario,
@@ -154,22 +229,36 @@ module.exports = {
   },
   async delete(req, res) {
     const { codigo } = req.params;
-
-    await Usuario.destroy({
-        where: {
-            codigo
-        }
-    }).then(retorno => {
-        if(retorno){
-            return res.status(200).json({
-                success: 'Usuario - excluido com sucesso'
+    
+    await Usuario.findByPk(codigo)
+    .then(async usuario => {
+        if(usuario){
+            await Usuario.destroy({
+                where: {
+                    codigo
+                }
+            }).then(async retorno => {
+                if(retorno){
+                    await Endereco.destroy({
+                        where: {
+                            codigo: usuario.idEndereco
+                        }
+                    })
+                    .then(() => {
+                        return res.status(200).json({
+                            success: 'Usuario - excluido com sucesso'
+                        });
+                    })
+                }else{
+                    return res.status(400).send();
+                }
+            })
+            .catch( error => {	
+                return res.status(500).json(error);	
             });
         }else{
             return res.status(400).send();
         }
-    })
-    .catch( error => {	
-        return res.status(500).json(error);	
     });
 
   }
